chore: move __test__ folder outside src
